+++ lib/hooks/useAnalytics.js
@@ -1,11 +1,12 @@
 import ReactGA from 'react-ga';
 import storage from 'local-storage-fallback';
+import Cookies from 'universal-cookie';
 
 export const useAnalytics = () => {
   return {
     init: (trackingId) => {
       console.log('Initialising ReactGA with trackingId:', trackingId);
-      ReactGA.initialize(trackingId, { debug: true });
+      ReactGA.initialize(trackingId, { debug: false });
     },
     trackPageViewed: (path) => {
       if (path) {
@@ -34,6 +35,24 @@ export const useAnalytics = () => {
         window.location.pathname + window.location.search,
         fields
       );
+    },
+    trackMailChimpParams: (query) => {
+      const cookies = new Cookies();
+
+      let campaignId = query.mc_cid;
+      let emailId = query.mc_eid;
+
+      if (campaignId && emailId) {
+        let cookieValue = {
+          campaignId: campaignId,
+          emailId: emailId,
+        };
+        cookies.set('nc_mct', JSON.stringify(cookieValue), {
+          path: '/',
+          maxAge: 3600, // Expires after 1hr
+          sameSite: true,
+        });
+      }
     },
     logReadingHistory: () => {
       let history = JSON.parse(storage.getItem('TNCHistory'));
