+++ lib/utils.js
@@ -6,7 +6,6 @@ import ListNode from '../components/nodes/ListNode.js';
 import TextNode from '../components/nodes/TextNode.js';
 import ImageWithTextAd from '../components/ads/ImageWithTextAd.js';
 import { format } from 'date-fns';
-import Cookies from 'universal-cookie';
 
 const ORG_SLUG = process.env.ORG_SLUG;
 const HASURA_API_URL = process.env.HASURA_API_URL;
@@ -119,25 +118,6 @@ export const renderAuthors = (article) => {
   });
 
   return output;
-};
-
-export const trackMailChimpParams = (query) => {
-  const cookies = new Cookies();
-
-  let campaignId = query.mc_cid;
-  let emailId = query.mc_eid;
-
-  if (campaignId && emailId) {
-    let cookieValue = {
-      campaignId: campaignId,
-      emailId: emailId,
-    };
-    cookies.set('nc_mct', JSON.stringify(cookieValue), {
-      path: '/',
-      maxAge: 3600, // Expires after 1hr
-      sameSite: true,
-    });
-  }
 };
 
 export const renderDate = (isoDate, renderTime = true) => {
